Move UpdateCrew to separate file.
